% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.TipologiaRodizio.R
\name{as.data.frame.TipologiaRodizio}
\alias{as.data.frame.TipologiaRodizio}
\title{Metodo S3 que retorna a representacao rm `data.frame` de um objeto da classe TipologiaRisco}
\usage{
\method{as.data.frame}{TipologiaRodizio}(x, ...)
}
\arguments{
\item{x}{objeto da classe `TipologiaRodizio`.}

\item{...}{eventuais argumentos adicionais.}
}
\description{
Este metodo simplesmente retorna o atributo `tabela` de um objeto da classe TipologiaRisco.
}
\examples{
\dontrun{
casosSuspeitos <- TipologiaRodizio(dados)
dfCasosSuspeitos <- as.data.frame.TipologiaRodizio(casosSuspeitos)
}
}
\author{
Bruno M. S. S. Melo
}
