% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TipologiaRodizio.R
\name{plot.TipologiaRodizio}
\alias{plot.TipologiaRodizio}
\title{Metodo S3 que plota na tela uma representacao visual do grafo do tipo `igraph` contido no objeto da classe `TipologiaRodizio`}
\usage{
\method{plot}{TipologiaRodizio}(x)
}
\description{
Metodo S3 que plota na tela uma representacao visual do grafo do tipo `igraph` contido no objeto da classe `TipologiaRodizio`
}
\examples{
\dontrun{
casosSuspeitos <- rodizioIdentificaSituacoesSuspeitas(dados)
plot(casosSuspeitos)
}
}
\author{
Bruno M. S. S. Melo
}
