% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.TipologiaRodizio.R
\name{print.TipologiaRodizio}
\alias{print.TipologiaRodizio}
\title{Metodo S3 que imprime na tela um data.frame que representa um objeto da classe `TipologiaRodizio`}
\usage{
\method{print}{TipologiaRodizio}(x)
}
\description{
Metodo S3 que imprime na tela um data.frame que representa um objeto da classe `TipologiaRodizio`
}
\examples{
\dontrun{
casosSuspeitos <- rodizioIdentificaSituacoesSuspeitas(dados)
print(casosSuspeitos)
}
}
\author{
Bruno M. S. S. Melo
}
