% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TipologiaRodizioMetodologiaGrafo.R
\name{TipologiaRodizioMetodologiaGrafo}
\alias{TipologiaRodizioMetodologiaGrafo}
\title{Identifica potencias mercados de risco de praticas colusivas a partir de grafo de licitacoes}
\usage{
TipologiaRodizioMetodologiaGrafo(grLicitacoes)
}
\arguments{
\item{grLicitacoes}{objeto do tipo \code{igraph} contendo um grafo direcionado de vencedores e participantes
de licitacoes;}
}
\value{
objeto do tipo environment, contendo os seguintes objetos:
\itemize{
        \item \strong{cmMercados} objeto do tipo \code{community} contendo todos as comunidades (mercados) obtidas a partir do grafo \code{grLicitacoes};
        \item \strong{grMercadosRisco} grafo do tipo \code{igraph} contendo os mercados de risco extraidos do grafo \code{grLicitacoes};
        \item \strong{vcMercadosRisco} vetor do tipo \code{numeric} contendo os identificadores dos mercados considerados de risco;
        \item \strong{vcEmpresasRisco} vetor do tipo \code{numeric} contendo os identificadores dos mercados de risco a que pertencem as empresas
              consideradas suspeitas de praticarem acoes colusivas. As empresas sao identificadas pelo atributo \code{names}.
         }
}
\description{
Identifica potencias mercados de risco de praticas colusivas a partir de grafo de licitacoes
}
\examples{
\dontrun{
grafoLic <- TipologiaRodizioMetodologiaGrafo(grLicitacoes)
}
}
\seealso{
\code{igraph}
}
\author{
Bruno M. S. S. Melo
}
