% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.TipologiaRodizio.R
\name{plot.TipologiaRodizio}
\alias{plot.TipologiaRodizio}
\title{Metodo S3 que plota na tela uma representacao visual do grafo do tipo `igraph` contido no objeto da classe `TipologiaRodizio`}
\usage{
\method{plot}{TipologiaRodizio}(x, ...)
}
\arguments{
\item{x}{objeto da classe `TipologiaRodizio`.}

\item{...}{eventuais parametros adicionais.}
}
\description{
Os diferentes agrupamentos representam empresas suspeitas de praticarem alguma acao colusiva num determinado mercado. As arestas apontam na direcao de um perdedor para um vencedor de licitacao. Empresas sao sempre perdedoras sao representadas por quadrados cinzas.
}
\examples{
\dontrun{
casosSuspeitos <- TipologiaRodizio(dados)
plot(casosSuspeitos)
}
}
\author{
Bruno M. S. S. Melo
}
